requirements.txt
# Core Streamlit and web framework
streamlit>=1.29.0
streamlit-option-menu>=0.3.6

# Document processing
docling>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
markdown>=3.5.1

# Web crawling
crawlee>=0.2.0
aiohttp>=3.9.0
requests>=2.31.0
urllib3>=2.1.0

# Data processing
polars>=0.20.0
pydantic>=2.5.0

# Async and concurrency
anyio>=4.2.0
trio>=0.23.0

# Progress and UI
tqdm>=4.66.0
rich>=13.7.0

# Logging and configuration
loguru>=0.7.2
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
validators>=0.22.0
pathlib>=1.0.1
typing-extensions>=4.8.0